sensing: implement the phy_3d_sensor
This patch is to implement the phy_3d_sensor in sensing subsystem. It's a
wrapper layer of AGM sensor driver for sensing subsystem. It supports
the accelerometer and gyrometer now.
This has been tested with lsm6dso and lis2dw12 on hardware and bmi160 on
native_posix simulator. It can work fine while sensor is either in polling
mode or in data ready trigger mode. And set_sensitivity function is just
implemented following api definition and not tested yet.

Tested with:
$ west build -p -b native_posix samples/subsys/sensing/simple/ -t run

Signed-off-by: Zhang Lixu <lixu.zhang@intel.com>